add option to skip cluster deletion on failure

+++ tests/e2e/framework/framework.go
@@ -52,9 +52,10 @@ func New() *Framework {
 }
 
 type SetupOptions struct {
-	SkipWaitForClusterDeleted   bool
-	SkipWaitForNamespaceDeleted bool
-	SkipClusterCreation         bool
+	SkipWaitForClusterDeleted    bool
+	SkipWaitForNamespaceDeleted  bool
+	SkipClusterCreation          bool
+	SkipClusterDeletedWhenFailed bool
 }
 
 type SetupOption func(opts *SetupOptions)
@@ -74,6 +75,12 @@ func WithSkipWaitForNamespaceDeleted() SetupOption {
 func WithSkipClusterCreation() SetupOption {
 	return func(opts *SetupOptions) {
 		opts.SkipClusterCreation = true
+	}
+}
+
+func WithSkipClusterDeletionWhenFailed() SetupOption {
+	return func(opts *SetupOptions) {
+		opts.SkipClusterDeletedWhenFailed = true
 	}
 }
 
@@ -108,7 +115,9 @@ func (f *Framework) Setup(opts ...SetupOption) {
 		f.Must(f.Client.Create(ctx, f.Namespace))
 
 		ginkgo.DeferCleanup(func(ctx context.Context) {
-			if !ginkgo.CurrentSpecReport().Failed() {
+			if options.SkipClusterDeletedWhenFailed && ginkgo.CurrentSpecReport().Failed() {
+				ginkgo.By(fmt.Sprintf("Case failed, skip cluster deletion: %s", f.Cluster.Name))
+			} else {
 				ginkgo.By(fmt.Sprintf("Delete the namespace %s", f.Namespace.Name))
 				f.Must(f.Client.Delete(ctx, f.Namespace))
 
@@ -116,8 +125,6 @@ func (f *Framework) Setup(opts ...SetupOption) {
 					ginkgo.By(fmt.Sprintf("Ensure the namespace %s can be deleted", f.Namespace.Name))
 					f.Must(waiter.WaitForObjectDeleted(ctx, f.Client, f.Namespace, waiter.LongTaskTimeout))
 				}
-			} else {
-				ginkgo.By(fmt.Sprintf("Case failed, skip deleting namespace: %s", f.Namespace.Name))
 			}
 		})
 	})
@@ -129,7 +136,9 @@ func (f *Framework) Setup(opts ...SetupOption) {
 			f.Must(f.Client.Create(ctx, f.Cluster))
 
 			ginkgo.DeferCleanup(func(ctx context.Context) {
-				if !ginkgo.CurrentSpecReport().Failed() {
+				if options.SkipClusterDeletedWhenFailed && ginkgo.CurrentSpecReport().Failed() {
+					ginkgo.By(fmt.Sprintf("Case failed, skip cluster deletion: %s", f.Cluster.Name))
+				} else {
 					ginkgo.By(fmt.Sprintf("Delete the cluster: %s", f.Cluster.Name))
 					f.Must(f.Client.Delete(ctx, f.Cluster))
 
@@ -137,8 +146,6 @@ func (f *Framework) Setup(opts ...SetupOption) {
 						ginkgo.By(fmt.Sprintf("Ensure the cluster: %s can be deleted", f.Cluster.Name))
 						f.Must(waiter.WaitForObjectDeleted(ctx, f.Client, f.Cluster, waiter.LongTaskTimeout))
 					}
-				} else {
-					ginkgo.By(fmt.Sprintf("Case failed, skip deleting cluster: %s", f.Cluster.Name))
 				}
 			})
 		})
